Refactor code to not use filter.
We were originally using the `fjgwpp_add_entries_filter` to add the videos
because it was part of the flickr-justified-gallery plugin. Since this
filter was removed, we modified the code and used apply_filters to run
the filter ourselves. Rather than have the filter it makes more sense to
run our code directly, which is what this refactor does.

See #9